Swap hash/uint256 test_random impls
